object fix and immediate update
